Refactor Gradio UI and response generation for LinkedIn and Tweet posts with enhanced engagement scoring and targeted questions
